vol, volser, and viced type fixes
Correct many uses of incorrect types for file offsets, sizes, etc. in
the volume server, fileserver, and the volume package.

Thanks to Tom Keiser who pointed out the incorrect types and suggested
solutions.

Reviewed-on: http://gerrit.openafs.org/1920
Tested-by: Russ Allbery <rra@stanford.edu>

+++ src/vol/ihandle.c
@@ -676,7 +676,7 @@ stream_write(void *ptr, afs_fsize_t size, afs_fsize_t nitems,
 int
 stream_seek(StreamHandle_t * streamP, afs_foff_t offset, int whence)
 {
-    int rc;
+    ssize_t rc;
     int retval = 0;
 
     if (streamP->str_direction == STREAM_DIRECTION_WRITE
@@ -703,7 +703,7 @@ stream_seek(StreamHandle_t * streamP, afs_foff_t offset, int whence)
 int
 stream_flush(StreamHandle_t * streamP)
 {
-    int rc;
+    ssize_t rc;
     int retval = 0;
 
     if (streamP->str_direction == STREAM_DIRECTION_WRITE
@@ -725,7 +725,7 @@ stream_flush(StreamHandle_t * streamP)
 int
 stream_close(StreamHandle_t * streamP, int reallyClose)
 {
-    int rc;
+    ssize_t rc;
     int retval = 0;
 
     assert(streamP != NULL);
